Update image versions
eirini commit SHA: 69cc40ccd3582dae646876c86e7b227c816bf556
eirini commit message:
Increase the timeout for the webhook test

because if the test runs before the binary has the time to
bind the port, the test will not find it listening there.

If the webhook secret is not there (e.g. as it happens if we
have run the tests before), then binding the port takes more time.

Co-authored-by: Dimitris Karakasilis <dkarakasilis@suse.com>