Update image versions
eirini commit SHA: f5f48da417a4e114fa22212b631594ed5a3ed35c
eirini commit message:
Fix canonical package name of the bbolt package

This is a very indirect dependency, via eirinix, cobra, viper, etcd. The
initial problem is in etcd-v3.3, with an incorrect import of
github.com/coreos/bbolt. This is fixed in etcd-v3.4, however, etcd-v3.4
is not a correct go module, and cannot be included in a go mod project!
So everyone must wait for etcd-v3.5 to be released with the fix. See
https://github.com/etcd-io/etcd/issues/11154#issuecomment-620886957.

Until then, we can rename bbolt. Fixing to v1.3.5 is unfortunately
required by go.mod's syntax. It should be upgraded if bbolt releases
again.