Change imagePullPolicy to IfNotPresent
This way we will optimize deploying eirini via usagea of cached images.
This is not dangerous since eirini uses unique sha256 versions of its
images. Another useful benefit is easier patching with locally built dev
images of the eirini components. Instead of pushing we can just load
images into a kind cluster and they are ready to be used without waiting
for pushes and pulls.

[#175269118]