Update image versions
eirini commit SHA: 3b0ad037b75b18313544e2742a4c8d02f2d5f8c8
eirini commit message:
Use PVC for the go cache volume when running on kind

Go cache is very sensitive to source code location and therefore cannot
be shared with the host. Instead, we are creeating a persistent volume
for the cache and copy the eirini code in the same directory on every
run. Thus first run is slow (as the cache is empty), but subsequent ones
are quick (as the cache is reused)

[#176012283]