Reconcile accidentally unresolved conflit
We had some files left in the deploy folder, containing changes that
hadn't made their way to helm/templates. In particular, we were missing
the instance-index-env-injector-job. We ported those changes to
helm/templates and deleted the deploy folder.

We think one of the commits that went wront is f2cc45a9.

Co-authored-by: Mario Nitchev <marionitchev@gmail.com>