Update image versions
eirini commit SHA: 90a1d2a47ca87b4a1922f133bc62e2a631a32ef6
eirini commit message:
Update counter annotation using jsonpatch with test

To ensure an 'atomic' increment of the counter annotation, we revert to
using jsonpatch which contains a 'test' operation. We ensure the
previous value is 1 below the value we are trying to set, or 'null' if
setting for the first time.

This way, we take advantage of patch being able to ignore changes to
other parts of the object it is patching, but make the patch fail if we
are clearly operating on stale data.

[#177044214]

Co-authored-by: Kieron Browne <kbrowne@vmware.com>