Reuse the EiriniX webhook name
This is the easiest way to address the cf-for-k8s upgrade issue of
clashing webhooks.

In short, for some reason kapp (the tool that deploys cf-for-k8s) does
not end up managing EiriniX webhooks. Thus when we decided to delete the
EiriniX hook from our template, after bumping Eirini on cf-for-k8s
results into the old webhook still being around. As the service it
references is no longer around, the webhook fails and prevents pods for
the LRP statefulset fom being created.
When reusing the webhook name, we just change the existing hook to be a
non-EiriniX one. As a side effect, kapp adopts the webhook which means
that it can be renamed/deleted in future versions.

More details here: https://cloudfoundry.slack.com/archives/CH9LF6V1P/p1614099287007700

Co-authored-by: Danail Branekov <danailster@gmail.com>