Update image versions
eirini commit SHA: 43b0cffdde9ac213519ee4e643c8fe87d2ca34f2
eirini commit message:
Set OwnerRef on Secret to StatefulSet

When creating a secret for a private registry image pull secret, set the
reference on the secret to the statefulset so that it is automatically
garbage collected when the stateful set is deleted.

Also migrate statefulsets of existing LRPs to adopt private registry
secrets.

[#177001209]

Co-authored-by: Danail Branekov <danailster@gmail.com>
Co-authored-by: Kieron Browne <kbrowne@vmware.com>
Co-authored-by: Mario Nitchev <marionitchev@gmail.com>